Rename migration config to reset it to false in all instances
We've been having a lot of bug reports of copycats losing their textures due to this option being true, probably because they have this option defaulted to true when they update to 1.3.x. Seeing that the migration process doesn't actually cause much lag, it's better to reset this to false to stop further reports.